Clustering simple approach
This was the simple approach I used to check the clustering part again it all runs fine. is nothing super crazy and is very primitive, but it runs with no issue.

Also the way I cleaned the data here is quite agressive since I wanted to see the possible results.

And also no plots, metrics or labels were saved atm so it's a must to run it again and save everything, the good thing is that it will probably not raise any error in the models since I already ran all of them without issues.